Refactor homogeneous to cartesian coordinates conversion
Explanation: when the w coordinate is zero, the vector represents a
direction, and therefore does not need to be divided by w. For instance,
a +90 degrees rotation of (1, 0, 0, 0) around the Z-axis generates the
vector (0, 1, 0, 0). The division by zero would lead to incorrect
values, but more importantly, the division would be meaningless, since
the zero w coordinate means that we are only considering the direction
of the vector instead of a specific point.